Fix ninja not being found
Throws an error if a binary is not found, which breaks the build.
Use command -v as it is safer.
Also, switch from ?= to := as it was causing ninja to not be found even with command -v.
We'll print what binary the user is using so that there is no confusion either.

Change-Id: I32a98b5bd9fce761e5c5f5122c7c377e3f43fff4
Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>
Signed-off-by: AKabhishek <abhishekkaushik.star@gmail.com>